# Practicas/practica3/practica3-rest/openapi.yaml
openapi: 3.0.3
info:
  title: API de Usuarios
  version: 1.0.0
  description: Servicio REST de la Practica 3 de COM-600
servers:
  - url: http://localhost:3002/v1
paths:
  /usuarios:
    get:
      summary: Lista usuarios con paginacion
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
        - name: limite
          in: query
          schema:
            type: integer
        - name: edadMin
          in: query
          schema:
            type: integer
        - name: ordenPor
          in: query
          schema:
            type: string
        - name: orden
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Pagina de usuarios
    post:
      summary: Crea un usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                correo:
                  type: string
                edad:
                  type: integer
      responses:
        "201":
          description: Creado, con cabecera Location
        "400":
          description: Datos invalidos
        "409":
          description: Correo ya registrado
  /usuarios/{id}:
    get:
      summary: Obtiene un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usuario encontrado
        "404":
          description: No encontrado
    put:
      summary: Actualiza un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                correo:
                  type: string
                edad:
                  type: integer
      responses:
        "200":
          description: Usuario actualizado
        "404":
          description: No encontrado
    delete:
      summary: Elimina un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
        "404":
          description: No encontrado